pyproject.toml
[tool.poetry]

name = "meddra_toolkit"
version = "0.0"
description = "Meddra toolkit to handle ontology files"

license = "MIT"

authors = ["Pierre Luce <pierre.luce@posos.fr>"]

readme = "README.md"

homepage = "https://pypi.org/project/meddra_toolkit"
documentation = "https://meddra_toolkit.readthedocs.io"
repository = "https://github.com/posos-tech/meddra-toolkit"

keywords = [
]
classifiers = [
    # TODO: update this list to match your application: https://pypi.org/pypi?%3Aaction=list_classifiers
    "Development Status :: 1 - Planning",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
]

[tool.poetry.dependencies]

python = "^3.9"

# TODO: Remove these and add your library's requirements
click = "^7.0"
minilog = "^2.0"
pydantic = "*"

[tool.poetry.dev-dependencies]

# Formatters
black = "=20.8b1"
isort = "=5.5.1"

# Linters
mypy = "*"
pydocstyle = "*"
pylint = "~2.6.0"

# Testing
pytest = "^5.3.2"
pytest-cov = "*"
pytest-describe = { git = "https://github.com/pytest-dev/pytest-describe", rev = "453aa9045b265e313f356f1492d8991c02a6aea6" } # use 2.0 when released
pytest-expecter = "^2.1"
pytest-random = "*"
freezegun = "*"

# Reports
coveragespace = "^4.0"

# Documentation
mkdocs = "~1.0"
pygments = "^2.5.2"

# Tooling
pyinstaller = "*"
sniffer = "*"
MacFSEvents = { version = "*", platform = "darwin" }
pync = { version = "*", platform = "darwin" }
ipython = "^7.12.0"

[tool.poetry.scripts]

meddra-toolkit = "meddra_toolkit.cli:main"

[tool.black]

target-version = ["py36", "py37", "py38"]
skip-string-normalization = true

[build-system]

requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
